+++ src/features/sicp/TableOfContentsHelper.ts
@@ -10,9 +10,6 @@ type TocNavigation = Record<string, TocNavigationNode>;
  * Returns the section slug that follows `section` in the textbook, per the supplied
  * toc-navigation JSON. Returns `undefined` if `section` is not present, or if the
  * entry has no `next`.
- *
- * The toc-navigation JSON is passed in by the caller so that the same helper can
- * serve both SICP JS (`toc-navigation.json`) and SICPy (`toc-navigation-py.json`).
  */
 export const getNext = (tocNavigation: TocNavigation, section: string): string | undefined => {
   return tocNavigation[section]?.next;
